Second fix for #2082801 - errors when converting ulongint expressions to floating-point (note: previous fix was actually for float to ulongint - careless old me...)
This was happening because for expressions (which were kept in registers), the sign bit check (used to work out whether to add 2^64 to the number, because fild treats values as signed) was apparently being done on the wrong register.  Hopefully fixed that.  Also renamed the variable for hoped increase in clarity.

Also, hexed the HPUSH'ed numbers for clarity/consistency.

Unit test added, fix for possible bug in f2ll unit test.

Note: both #2082801 fixes have been done on both x86 and SSE emitters.  The fixes were the same.

r5270