fbportio and gfxdata: Separate out of runtime build
The fbportio build has always been separate, for the gfxdata
stuff this is new. For both cases we're much better off putting
the generated code into the source, instead of rebuilding it
as part of the runtime build process, because they're almost
never changed.

Besides that the runtime build becomes much, much easier, since
there no longer is any code that needs to be generated before
it can be compiled, and the makefile doesn't need to worry about
a native build toolchain in case of cross-compiling.

Both runtime/gfxdata/ and runtime/fbportio/ get their own small
makefiles allowing to regenerate the headers when needed.