One runtime per build
It's easier to build only one runtime at a time, instead of one for each
enabled target. This wouldn't be needed often so it only adds unnecessary
build time and bloat to the makefile.

A multi-target compiler is still easily possible by using ENABLE_TARGET_*
defines, while TARGET selects the default target. And the corresponding
runtimes need to be produced by a separate build, which is acceptable
since this is a rather rare case. The multi-target compiler should be
flexible enough to allow extra runtime libs to be installed and used at
any time.

At least a modular multi-lib toolchain similar to gcc-multilib is very
easy to achieve, i.e. 32bit vs 64bit cross-compiling. fbc and gcc/ld used
by it use either lib32 or lib64.

And the fail-safe choice, similar to gcc, is to build a pure cross
toolchain, with bin/<target>-fbc and lib/<target>-freebasic/.