Use explicit new/ build directory inside source tree
We need some place to put compiled objects etc., and previously we had
multiple obj/ subdirs. With the combined toplevel makefile it seems to be
a good idea to "reserve" one toplevel directory to be the build directory.
This best avoids having objects mixed/hidden in the source tree somewhere,
and allows for multiple builds in a single source tree, by simply using
a different build subdir each time.

This might even be better than having a completely separate build
directory (outside the source tree) as is common with autoconf, because
it is never necessary to cd away from the source toplevel. It should still
be possible to do parallel builds by launching two makes with different
build dirs, so all this achieves pretty much the same effect (plus more)
of what autoconf can do.

A nice extra of using a new/ subdir instead of multiple obj/ subdirs is
that the compiler can be built into new/ as if it were installed there,
so that it can be run and tested without being installed anywhere (as long
as the prefix isn't hard-coded).