Reallow standalone builds, now meaning that fbc is in . instead of ./bin
This new kind of "standalone" build only changes how fbc expects to be
located relatively to the bin/, include/ and lib/ sub-directories.
Unlike previously, this is now completely separate from the path prefix
thing (the question of whether the compiler is relocatable or not).
This is needed for self-contained Windows/DOS installations, when only fbc
is wanted in the PATH, but not the binutils etc. in bin/.

New makefile option: ENABLE_STANDALONE