Don't expect target subdirs in bin/ and lib/ (switching to GNU triplets)
To be compatible to binutils/gcc cross-compiling toolchains, the GNU
triplets will be used instead to distinguish duplicate tools/libs.
The triplets sure aren't as pretty as FB's shorter target names, but it's
worth it, to support cross-compiling with FB on Linux systems -- there
it's all based around the triplets.