Use the new ENABLE_<TARGET> defines/triplets in the compiler
Each target has a default triplet that fbc.triplet will be set to when the
target is loaded. It should also be possible to override that with a
triplet directly passed to -target (but that's not yet implemented).

The triplet must be used as prefix for binutils (instead of looking for
them in target-specific subdirs).