Update the readme
Besides another rewrite of the installing/running section (when will it
sound good *and* be useful *and* small?), I'm adding Sterling Christensen
and myself to the dev list. Some of the key points that everyone worked
on are back in the list too (similar to old readme, but updated), maybe
it *is* ok to have that, feel free to update as you see fit.