makefile: Don't build with -g by default
The default should be close to the release build, so that it's getting
the most testing. Also the huge libs are annoying and not needed that
often.

Using -exx by default on the compiler would be more useful than -g in many
cases, but -exx also is a huge slowdown and can cover up bugs, so that one
cannot be on by default.