makefile: Treat non-working uname (e.g. no Msys) as sign of Windows/DOS
This is the first requirement for allowing to build without MSYS installed,
which would really be nice to have.

uname doesn't exist on Windows/DOS normally, but on all (?) other systems
we support. But this will now also check the value of COMSPEC, to find
out whether it's DOS or Windows. This should also catch the case of a
non-Windows/DOS system without uname, assuming it wouldn't have such a
COMSPEC value.