makefile: Avoid using $(or) and $(and) for backwards compatibility
Older make versions again... e.g. DJGPP's make, they don't have these
functions.

This means the triplet parsing code can't use the neat functional approach,
and the code needs to be duplicated for HOST and TARGET, but oh well.
This is not that bad, it allows to make extra decisions or parse triplets
more exactly.