makefile: Always use /usr/local prefix by default
Currently installing into '.' isn't a good idea, it would overwrite the
headers (and maybe other things in the sources), and it'd most likely be
bad to assume something like /mingw or c:/mingw will work for the Windows
build.

Since at the moment 'make' already results in a usable installation in
new/, it doesn't matter much for development what 'make install' does.