manifests: Still useful, but remove the scripts made for old svn
The scripts were about automatically generating the manifests based on
file name patterns, this was needed because in the old directory layout
source/build directory was the same; now with the new/ build subdirectory
it's easy enough to make releases based on that. Manifests can then be
made by removing the objects from the build directory, then running 'find'
and 'sed' on the build directory (or whereever the release is being put
together). Same goes for the manifest-based zipping.