Add fb_program_icon stub to libfbgfx, replacing $$fb_icon$$.o
When given an .xpm, fbc compiles it into an object that holds the
"fb_program_icon" global pointer variable, allowing the gfxlib to
access the xpm data at runtime.

$$fb_icon$$.o was used when the user didn't give an .xpm,
to provide a null fb_program_icon pointer to the gfxlib.

Now this stub is contained in the gfxlib itself, so $$fb_icon$$.o is
no longer created on every single link.