When cross-compiling, ask the cross-gcc for libs
We don't know the path to the gcc libs (they're in gcc's private
target/version-specific subdir), so we need to do "gcc -print-file-name".
It's the same for cross-compiling, only that the "cross-gcc" needs to
be asked.

Standalone builds continue expecting the cross libs to be in the
corresponding <target>-lib/ subdir. Non-standalone builds on the
other hand can't expect the cross libs to be present in <target>-lib/,
that would defeat the point of installing fbc "into the system".