Revert "Allow fbc to use programs from PATH on all systems, [...]"
This reverts commit d765f8d67b21910f6fbda1b20a5613e3e1ae4fd7.

The idea was to allow running the Win32 fbc without having binutils/gcc
in the same location, as is done on Linux already. It would have been
very useful for development (compiling without installing).

However, fbc launches as/ld etc. using exec(), and doing exec("ld", ...)
only works on Linux. On Windows, it does only a CreateProcess(), so it
doesn't invoke the ld from PATH, but it will look for ld.exe in the
current directory.

Using shell() could solve this, but shell() would require us to be careful
about shell syntax and quote special chars (not just spaces) everytime
when calling as/ld, that's not worth it.