Fix #3404250: Pass input .a files on to ld as-is, and inside -( -)
Note: Libs given by file name like that (instead of through the -l option)
are not added to objinfo at the moment, but that shouldn't hurt since
usually -l or #inclib are used (those are what objinfo is there for).
.o files aren't added to objinfo either, even if they're given together
with a .bas and -c.

The libs must be passed to ld behind any user objects, or its content may
not be used, causing undefined references. And then it's also best to just
include it in -( -) too, which will work out inter-lib references.