Fix subtle 7d4d018 regression: broken QB forward CALLs
In that symbAddPrototype() call the last parameter isn't given, it is
optional so that's ok, but removing the libname parameter caused everything
to shift without causing an error or even a warning since all following
arguments are plain zero. I should have thought of that when removing a
parameter on a function with optional parameters...

Luckily only this one line is affected, and the compiler doesn't use
forward CALLs, so it's safe to just rebuild.

(The qb/call.bas log test failed with undefined references because of this)