Remove compile-time TARGET_X86[_64] defines, this must be done at runtime
For 64bit support later, some way must be found for the compiler to
use the proper sizes for longs and pointers, depending on the target
OS/arch selected by the user at runtime.

Turning FB_LONGSIZE etc. into variables is impossible at the moment,
since it's used in global var initializers etc, but eventually it must
be done. Perhaps long could be remapped as it's done with integer under
-lang qb? As long as it doesn't change overloading choices...