Make the temporary dos ld args file respect the path given with -x
Previously it was only based on the path of the main module, however it
makes sense for it to respect -x, since the ld args file is a temporary
file involved in creating the executable, not compiling the main module.
By default, -x respects -m automatically, too, so there is no change
except if -x is used.