Fix null op optimizer to catch cases where null ops propagate up
This changes hOptNullOp() to try and eliminate children first, before
checking the current node. That lets it eliminate null ops recursively
from ground up.

Closes: #3434928

In this case the null op optimizer didn't catch everything it could,
leaving the tree open for further optimizations. This broke the assumptions
on the toplevel BOP being final (to be used for setting x86 flags) in
astUpdComp2Branch().

It relies on the BOP to be final, but the BOP was optimized out during the
following astAdd(), resulting in a conditional jump without the expected
instruction to set the flags in front of it.