Use outer scope for SELECT's temp var and destroy it at scope breaks
Closes: #3446210

SELECT CASE copies the expression to a temporary variable, to hold it for
all the CASE comparisons. For a string expression it's a string, for
wstring expressions it's a wchar ptr that's set to a newly dynamically
allocated buffer (since there are no proper dynamic wstrings yet).

In order not to leak memory, both types of temp strings must be cleaned up.
So far this was done (only) at END SELECT, thus leaking the strings if the
END SELECT was not reached, for example due to an EXIT FUNCTION from a
CASE block.

This patch...

* causes the FB_SYMBATTRIB_TEMP flag to be removed from the string,
  giving symbGetVarHasDtor() a chance to return TRUE for the variable if
  it has a dtor (i.e. the string will be destroyed at astScopeBreak()s etc)
  This trick is stolen from FOR loops which do the same (probably because
  of UDT iterators though, which aren't currently allowed in SELECT).
  Perhaps the same could be done by directly using symbAddVar() instead of
  symbAddTempVar(), but that would probably be more code when having to
  provide all the parameters.

* adds a new FB_SYMBSTATS_WSTRING flag and uses it to mark the temporary
  wchar ptr of a 'SELECT CASE wstring'. A new check in symbGetVarHasDtor()
  will recognize it and return TRUE, letting astBuildVarDtorCall() destroy
  the wstring with the same code used before at END SELECT.

* adds an outer scope around the whole SELECT ... END SELECT compound,
  in order to have the temp string destroyed at END SELECT and not later.
  Scoping the temp also allows its stack space to be reused. This trick
  is again stolen from FOR loops.

All in all the SELECT CASE temporary is now handled by scopes, and not
manually anymore, even including the wchar ptr thanks to a "dynamic
wstring" hack.