Unix rtlib: Temporarily reset the console during dylibload/free()
... in order to let the shared lib get to see the original
terminal state, because if it's an FB lib it will have a 2nd
rtlib that restores the terminal state on exit, when the shared
lib is unloaded. That could be *before* the program's main rtlib
(dylibfree()) or *afterwards* at process exit. Since the order
can't be predicted any "nested" rtlibs must know the proper
original state, in case they shutdown last.

The same thing is already done for shell() and exec() functions,
probably for the same reason (see 1be51c1).

This should fix the reported console issues with FB shared libs,
but it should be noted that in general this rtlib issue still
exists, for example if the system's dlopen() is used directly,
instead of FB's dylibload() wrapper. It also shows up in case
an FB program is shutdown due to an stack overflow, there's not
much that can be done about that though.