makefile: Combine HOST/TARGET variables into just TARGET
The main reason for this is that HOST is already a used environment var
name on OpenSUSE (and perhaps other systems), and since make inherits any
environment vars into the makefile we would have to use a different name.

The better solution however is to just use a single TARGET variable, so
cross-compiling FB can be done via 'make TARGET=...' similar to
'fbc -target ...'.

Having the separate HOST/TARGET allowed to build compiler/runtime for
different targets at the same time, but since TARGET no longer has an
effect on fbc (its default target always is its host), there is no use
in doing that. It's still possible to do this:
    make compiler TARGET=a
    make rtlib TARGET=b
to get the same effect.