gengas: Fix POPing to byte/short (closes SWAP bug #3215431)
POP with byte/short destination had "eax" hard-coded, without
handling cases when eax was used already.

An example of such a case is POPing to an array element
destination with a variable used as index in the array access.
For such an array access, the variable is loaded into a register,
which is included into the stack offset calculation. In this case
"eax" was used, which POP's byte/short handling code trashed
without checking, thus breaking the array access it was doing
later to store the byte/short to its destination in the array.

POP to byte/short is currently only used by the PUSH/POP code
generated for -gen gas SWAP, and this problem caused SWAP on
byte/short array elements to crash.