Fix SWAP on bitfields to work with -gen gcc and use less temp vars
The code path for SWAP on bitfields seems to have been broken for -gen gcc
for a while, as it would do a POP without ever having done a PUSH...
besides PUSH/POP not being supported by the C backend at all.

This whole piece of code seems to be unnecessarily complex, because the
paths that handle bitfields used temp vars and PUSH from or POP to those,
i.e. it tried to integrate proper bitfield accesses for the lhs or rhs,
or even both, with the normal PUSH/POP code. The C backend part on the
other hand mixes things up because it uses a temp var instead of PUSH/POP,
so in theory if it would have been working it would have used 3 temp vars.

I changed it to try PUSH/POP only for -gen gas and if there are no
bitfields involved, and otherwise (C backend or bitfields) fall back to
using one temp var.