rtlib: Add "real" Rnd() algorithm using Win32 Crypto API or /dev/urandom
New algorithm number: 5
Implemented by MOD. The goal is to provide much better random numbers.

It's not 100% clear whether the CryptAcquireContext() & co functions are
guaranteed to be present on old Windows systems, but testing on a
Windows 95 (version 4.00.1111) didn't show any issues with loading and
running the exe. It seems though that the Crypt*() functions don't always
work properly, in which case the new Rnd() algorithm will fall back on to
the existing mersenne twister algorithm. (Same for the Linux /dev/urandom
version, if there are any failures)