gfxlib2: Fix lock up with GET_WINDOW_POS on minimized window
When doing ScreenControl(GET_WINDOW_POS) while the graphics
window was minimized, the GetWindowRect() would return a
position of -32000, -32000 (the place where minimized windows
are hidden) and then do a SetWindowPos() with those values,
causing the window to be "re-minimized" immediately. With the
FB program doing such ScreenControl()s in a loop, there's a
race condition there, but if unlucky it could prevent the
window from being able to be restored to normal.

Now GET_WINDOW_POS won't do SetWindowPos() at all anymore,
so the lock up won't happen anymore. Note that it still can
return -32000, -32000 to the FB program as before, if the
graphics window is minimized. IsIconic() could be used to check
for that situation, but so far it seems better not to change
the behaviour there.