C backend: Fix fbc crash on circular dependency of UDT/funcptr
See also: tests/pointers/funptr_fwdbyref.bas

In this case the proc param was 'byref as UDT', i.e. emitted as pointer,
but hDtypeToStr() didn't get to see that because the "*" was added
manually. This prevented hEmitUDT() to resolve the circular dependency
by using a forward ref, that ran it into infinite recursion.