rtlib: Fix wstring Input buffer overflow and *wstringptr support
- Buffer overflow:
  fb_InputWstr() was passing the full buffer length (in wchars) to
  fb_FileInputNextTokenWstr(), which expects the length excluding
  the null terminator though.

- Dereferenced wstring pointers (user-allocated wstring buffer):
  Input'ing to a *wstringptr wasn't working at all since 0 (the
  length indicating a user-allocated wstring) was passed to
  fb_FileInputNextTokenWstr().

Now wstring Input uses a temporary buffer and the proper string
assignment function, like the normal zstring Input does. Because of
that it's now restricted by the same buffer length limit as the
normal zstring Input.

One possible change for the future is to remove the temporary
buffer for the {w|z}string Inputs, it's only really needed for the
string Input, which can only fill the string descriptor after
having read the data. However, Input is File I/O, so it's not the
most speed-critical function, and because of that using the
temporary buffer and the string assignment functions seems like
the cleaner solution.