C backend: Remove dupdecl mechanism
The dupdecl flag was used to mark built-ins aliased to the same function.

For example:
	hex alias "fb_HEX_i"(as integer) as string
	hex alias "fb_HEX_i"(as any ptr) as string    (dupdecl)

If both functions are used in the same module, the C backend would emit:
	string* fb_HEX_i(integer);
	string* fb_HEX_i(void*);

which isn't allowed in C. The dupdecl flag tried to solve this by telling
ir-hlc to not emit the hex(any ptr) overload.

However, 1) that still requires the hex(integer) overload to be used in
the same module, or otherwise there won't be any declaration of fb_HEX_i()
at all, and 2) in general this problem can happen with any FB functions
but there is no easy way for the C backend to deal with that.

Because of this it seems much better to remove the dupdecl flag and fix
up the few cases where fbc uses such declarations (hex/bin/oct).

Note: ir-hlc's FTOI builtins reused the dupdecl flag to avoid being
declared like normal rtlib functions. They are now using a new
IrHlcBuiltin flag instead.