ast: Remove astAdd() recursion
astTypeIniUpdate() and hCallCtors() now use LINKs instead of recursive
astAdd() calls. This means that astAdd() is no longer recursing, and
that the trees that were previously added recursively are now subject
to astAdd()'s updating and optimizations, as they should be.
(this was needed to unify the bitfield handling)

Besides that, hCallCtors() is now done from astProcEnd(), instead of
astAdd(), and it uses astAddAfter() to insert initialization code at
the top of constructors where needed. Doing this during astAdd() as
before would have required to LINK the ctor init code together with
whatever statement is being added, which would have worked aswell,
but isn't as pretty.