C backend: Enable nested FIELD removal to fix #3421351
In this scenario we had two nested field accesses on the lhs of an
ASSIGN. The first FIELD was properly used to detect the bitfield
assignment in astLoadASSIGN(), but the second FIELD was still in,
and was turned into a bitfield access by astLoadFIELD().

This meant there was a BOP on the lhs of an assignment, which doesn't
make sense. The lhs' vreg was an IR REG, i.e. irIsREG(v1) = TRUE,
which caused the C emitter to emit bad code.

The fix is to enable the nested-FIELD-removal optimization which
already existed for the ASM backend.