Add default ctor to UDTs if base or fields have *any* ctors, not just default ones
The CtorField UDT flag wasn't propagated up correctly, since it only
triggered for fields with default ctors, but not for fields with only
non-default ctors.

This wasn't an issue so far because such fields aren't allowed (the
parser will require at least an '= ANY' initializer on the field
declaration, and then the field initializer syntax itself triggers the
FieldInited flag which causes a ctor to be added, like the CtorField
flag would have done otherwise), however it's a problem for base UDTs
(which basically are just fields too).