Fix #3372520: Buggy type initialization through pointers (i.e. NEW)
TYPEINI_ASSIGN does an astNewDEREF() to assign to objects through pointers,
e.g. in the case of NEW, but sometimes the DEREF could end up having the
wrong datatype due to lack of a typeDeref(), causing the following ASSIGN
to fail.

Now it should do the typeDeref() when needed, i.e. when assigning "to" the
base symbol (the pointer in case of NEW). NEW passes AST_INIOPT_DODEREF
to astTypeIniFlush() to allow this to be detected.

The typeDeref() is not needed in case it's assigning to another symbol,
which seemed to always have the correct type already, e.g. an UDT field.
In that case the base symbol is the UDT pointer, while the TYPEINI_ASSIGN
assigns to a field symbol. The astNewDEREF() then uses the fields' type,
instead of deriving its type from the base symbol.

The solution feels pretty hackish but seems to work, I added some extra
assert()s to catch this issue in the future, in case the current check
isn't enough yet.

There is a certain connection to the cInitializer() parsing, which requires
NEW to pass FB_INIOPT_DODEREF to handle the "initialization through a
pointer" case, though this information isn't stored in the TYPEINI tree's
nodes, so it currently can't be used during TYPEINI flushing.

This commit also reverts c9535c309a8d177896b48ac34832e0bd5f42368d, which
seemingly tried to fix this same issue in a different and incomplete way.
These old changes don't seem to be needed anymore with the new fix.

This patch also fixes NEW issues with the C backend. The bad data type on
the DEREF caused gcc to complain about some of the generated C code; now
it works fine.