pp: Fix uninited ptr deref on variadic macro calls with empty vararg param
The argument table (containing the DZSTRING structures) that will hold
the arguments of a macro call is not cleared on allocation (it's a NOCLEAR
list), but only the used entries are cleared as soon as needed.

In case there was no arg given for the vararg param of a variadic macro,
and empty arg is assumed and the corresponding argtb entry must be cleared.
However a DZstrReset() was done instead of clearing this argtb entry,
causing "random" crashes depending on the uninitialized argtb memory.

This bug is the result of my previous testing implementation of variadic
macros which used static strings that required resetting, rather than
being dynamically allocated and requiring clearing as it is now.

This patch additionally removes the "doskip" mechanism from the
hLoadMacro[W]() functions. In case there are too few args specified,
the argtb entries for the remaining args will now be initialized (cleared)
instead, so it's no longer necessary to completely skip the macro
expansion in order to avoid using uninitialized argtb entries.