tools execution: Fallback to PATH for non-standalone on all systems
This allows the non-standalone build to fall back to using system-wide
tools (as, ar, ld, gcc...) on *all* systems, not just Unix.

That way it's possible to run a non-standalone fbc, even if it's not
installed into the system directory containing the tools (e.g. C:\MinGW),
provided the PATH environment is set up accordingly.

- On Unix, exec() will use the PATH, but on other systems (at least Win32),
  shell() has to be used