Set DONTINIT stat on SELECT CASE integer temp var
This causes SELECT CASE temp vars to be marked as DONTINIT unless they're
strings. This means integer temp vars will be treated as uninitialized,
which is true, because there is no DECL node created that would clear
them, and the following astNewVAR() only clears the temp var if it's a
string.

The effect is that the "branch crossing ..." warnings will be suppressed,
which should be fair: SELECT's integer temp vars won't be used anymore
once one of the CASE blocks has been reached, unlike string temp vars
which need destroying at scope end/breaks.