Treat SELECT CASE wstring temp as object with ctor
This turns the "branch crossing ..." warning into an error for code like
this:

	dim as wstring * 10 w

	goto label1:

	select case( w )
	case wstr( "a" )
		label1:
	end select

Jumping over the initialization of the hidden wstring should be an error
because it breaks the hidden clean up code. The variable and clean up code
is hidden and belongs to SELECT CASE, it is not visible to the user and it
is not written by the user.