Change -c, -r and -pp to complement each-other
This adds some guarantee to the outcome when using them together, which
doesn't really make sense, but nevertheless it's good to clarify the
behaviour.

-c no longer partially overwrites -r, and -r no longer overwrites -c,
instead they both enable their flags (those that are different from the
fbc default) and that's it. So now, using -r and -c plays out as -r,
because -r stops the compiler earlier.

-pp also won't overwrite -c/-R anymore, allowing -pp to be used in
combination with the -c mode, and using -pp together with -R will
happen to work, since -pp does the whole emitting step.

There are other odd cases such as -forcelang -lang that can partially
overwrite each other when used in the right order.