Fix #2131139: "SELECT CASE wstring" allowed integer expressions in CASE
"SELECT CASE wstring" uses a temp wstring, a fake dynamic string.
It's actually a wchar ptr variable for which a buffer is allocated,
assigned from the given wstring expression, and cleaned up at scope
breaks.

Because of this implementation detail, CASE statements have to do an
extra DEREF, to allow comparisons between the temp wcharptr and the
CASE expression, for example:

	*wcharptr = expression

Coincidentally, the compiler (more specifically astNewBOP()) allows this
form of comparison involving DEREF'd [z|w]string ptrs to work like string
indexing, to allow for easy access to string characters in the context of
comparisons. Normally *wcharptr results in a string expression, not a char
or wchar integer expression, but comparisons have this special case.

The fix changes astNewBOP() to check whether it's looking at a normal
*wcharptr expression, or that of a SELECT CASE temp var. Since commit
e310601d the fake dynamic wstring is marked with the special
FB_SYMBSTATS_WSTRING flag, which can now also be checked by astNewBOP().