SELECT CASE: Don't use temp var unless needed
This patch changes SELECT CASE to avoid using a temporary variable when
given simple expressions. For example:

	dim i as integer
	select case i
	...

in this case creating a temporary variable is unnecessary; and so SELECT
CASE will now just use the given variable directly for comparisons in
following CASE statements.

This can only be done when the given expression consists of just a plain
variable access. Any more complex expression (BOPs or even CALLs etc.)
should not be evaluated multiple times.

This is an optimization that results in better code, and it also provides
a work-around for "branch crossing ..." errors/warnings when GOTO jumping
over the declaration of SELECT CASE's temp var (if there is no temp var,
there is no problem with GOTO).