makefile: Build rtlib/gfxlib2 with -Werror-implicit-function-declaration
-Werror-implicit-function-declaration should turn these easily overlooked
warnings into errors that stop the FB build. The rtlib/gfxlib2 code has
been affected by this repeatedly, and it may happen again, when compiling
on less tested Unixy systems. For example, FB 0.24 contained rtlib fixes
to allow building it for x86 GNU/Linux from a x86_64 GNU/Linux system.
The difference in headers had caused some system functions to be left
undeclared, which is a serious bug, because the implicit declaration
that gcc will use is typically wrong.

This command line option is already mentioned in the gcc 2.95.3 docs [1],
so there shouldn't be an issue with slightly older gcc versions (unless
it's older than gcc 3).

[1] http://gcc.gnu.org/onlinedocs/gcc-2.95.3/gcc_2.html#SEC8