parser: Handle namespaced constants on lhs of assignment (closes #3549349)
Given code like this:

	namespace ns
		const foo = 1
	end namespace

	dim foo as integer

	ns.foo = 1

Starting at cStatement(), cProcCallOrAssign() parses the namespace prefix
"ns." and then calls hAssignOrCall() to parse and handle the member
identifier. This is typically necessary to call namespaced procedures
or access namespaced variables on the lhs of assignments.

However, hAssignOrCall() did not handle constants, ultimately causing
cProcCallOrAssign() to abort and return to cStatement(), despite having
parsed the namespace prefix already.

Then the following cAssignmentOrPtrCall() tried to parse the line, without
knowing about that skipped namespace prefix, causing it to treat the line
"ns.foo = 1" as if it was just "foo = 1".

(if there happens to be a variable using the same id as the constant, the
compiler just seemingly allowed a constant on the lhs of an assignment)

The fix seems to be: to treat constants much like variables during
hAssignOrCall(), even though there is no construct allowing them in that
place, to ensure they're getting parsed properly, and then letting the
cAssignment() (that hAssignOrCall() uses for variables) report the
"cannot modify constant" error.