gfxlib2 X11: Fix uninitialized access on maximize with DRIVER_NO_SWITCH
The "case ConfigureNotify" block would simulate ALT+ENTER in case the
window was maximized, and fall through to the "case KeyPress" block in
order to let it react to the ALT+ENTER.

However the "case KeyPress" had an extra "no DRIVER_NO_SWITCH" condition
that could prevent it from entering the ALT+ENTER handling code, and
instead enter the else block which assumes the X11 event to be a KeyPress
and not a ConfigureNotify. That could presumably lead to bad results,
when maximizing while DRIVER_NO_SWITCH is set.

Anyways, it seems like this case would not easily happen in practice,
because DRIVER_NO_SWITCH also tells X11 to not allow maximizing the
window. (a hint which I think window managers are not necessarily
required to follow)