ASM backend: Ensure to preserve EAX/EDX during unsigned MUL
The main problem was EDX not being preserved in a case where it wasn't
used by the dest vreg, but by something else.

Also, the old code section after the MUL (for storing the result),
didn't properly handle the case of eax not being preserved on stack
in case the source vreg was an immediate and eax was free.

1. In case of a dst like "[edx+...]" (i.e. it uses edx but it isn't edx),
and src being an immediate, src will be put into edx (thus "trashing" it),
and before that, edx and dst's value are backed up on stack because they're
still needed later.

2. Then dst's value must be put into eax (as 2nd MUL operand), which is
done by reading the dst value backup from stack. (since direct dst access
is no longer possible, because it depends on edx, and edx was "trashed").
The dst value is removed from stack, and assuming eax is free, there's no
need to add a backup of eax instead, so the stack is left with only the
edx backup (still needed later to store the result into dst).

3. After the MUL, this code section:

	if( eaxindest = FALSE ) then
		if( edxindest and (dvreg->typ <> IR_VREGTYPE_REG) ) then
			hPOP( "edx" )
			outp( "xchg edx, [esp]" )

would now corrupt the stack, because it assumes eax being there,
but it isn't.

It's not clear to me whether such a case (eax free and source being an
immediate) actually is a possible case that can happen with the ASM
emitter, but it's best to fix the bad code paths.

(Given that eax always (?) is the first reg used, and there are two vregs,
source and dest, what are the odds of eax being unused? ... perhaps with
vregs that don't use registers.)