ASM backend: Make CSNG(DOUBLE) truncate to SINGLE if needed
The problem is that for statements like

	double1 = cast( single, double2 )

fbc generated just a load and a store:

	fld qword ptr [ebp-...]
	fstp qword ptr [ebp-...]

and since the FPU stack (st(N) registers) won't truncate double (qword)
to single (dword) automagically, no cast to single would happen here.
It was equal to:

	double1 = double2

It seems necessary though that csng() converts the value to SINGLE,
after all it was explicitly specified. Cases like this:

	single1 = double2

do not have this problem because the load/store code generated for them
is enough to do the conversion:

	fld qword ptr [ebp-...]
	fstp dword ptr [ebp-...]

(note qword vs dword)

The way to get the conversion for csng() is to use a temp var on stack,
into which the SINGLE value is temporarily stored and read back into
the FPU stack (now possibly truncated). After that, any stores from the
FPU stack will use that truncated value and give the correct result.

This patch adds the new emitter operation CONVFD2FS (double to single)
which does the additional store to stack & read back from stack.
With that,

	double1 = cast( single, double2 )

results in the following sequence:

	fld qword ptr [esp+...]

	fstp dword ptr [esp+...]
	fld dword ptr [esp+...]

	fstp qword ptr [esp+...]

(note the additional store to stack & read back from stack)
This matches the code generated by GCC for C code like:

	double1 = (float)double2;

The additional code is generated from astLoadCONV() for conversions from
DOUBLE to SINGLE. However, this patch also adds some checks that prevent
the additional code from being generated in some cases where it's clear
that it's not going to be needed, such as:

	float1 = csng( double1 )

The DOUBLE to SINGLE truncation done by cast() is unnecessary here,
because the value will be truncated when stored into a single anyways.

This patch also includes a fix for the compile-time evaluation of DOUBLE
to SINGLE conversions; instead of doing nothing it will now truncate the
float value stored in "FBVALUE.float as double" using csng(). Note that
this relies on the fixed csng(), requiring the compiler to be rebuilt
with itself after this patch.