astCheckConst(): Remove duplicate/unnecessary warnings
astCheckASSIGN() and astCheckConst() now have a parameter telling them
whether or not to show warnings.

If there's an astNewASSIGN() more-or-less immediately following the
astCheckASSIGN() then we don't want the astCheckASSIGN() to show warnings,
because the astNewASSIGN() will take care of that.

In case of default values for function parameters the astCheckASSIGN()
is the only thing that's done, so the warnings should be shown then.