Remove "mcount" + some others from global namespace
Under BASIC mangling mcount becomes MCOUNT, so there is no collision
with the real mcount() function, so it may aswell be allowed.

It's true that there's a huge problem with code such as:

	sub rename alias "rename"( )
	end sub

or even

	extern "C"
		sub rename( )
		end sub
	end extern

because it shadows the system's rename() function without notice.

But having some of the affected symbols in the global namespace to
prevent such redefinition is useless, because *any* symbol can be
shadowed that way, for example malloc() or fwrite() or system() etc.

Overriding any symbol that way can cause lots of trouble, if we wanted to
make Extern "C" safer, we'd have to blacklist all CRT/system functions
used by the rtlib, which is too much work, not to mention that many are
re-#defined to implementation-specific names.

Case-preserving name mangling just has to be used with care.

	http://www.freebasic.net/forum/viewtopic.php?f=3&t=20386