ast: Fix faulty IDX optimization
The hOptConstIdxMult() pass searches for BOP( *, IDX, CONST ) expressions,
and if found, removes the BOP and instead puts the constant into the
ASTNODE.idx.mult field of the IDX node, allowing the ASM backend to
generate better code.

For example
	mov dword ptr [ebp+eax*4-M], 0
instead of
	imul eax, 4
	mov dword ptr [ebp+eax-M], 0

The multiplier on an array access can be included directly into the
address expression of an ASM instruction, but this works only for certain
multiplier values, and the ASM backend only supports this for multipliers
in the range 1..9.

However, the AST optimization only did a "if( c < 10 ) then" check,
which allowed negative multipliers to pass through unnoticed.

The ASM backend on the other hand uses a "if( mult > 1 ) then" check,
and thus it would just ignore the negative multiplier, causing the final
address calculation to be wrong.

The C backend was not seriously affected by this problem, because it never
ignores the multiplier (it is not limited by the C language as much as the
ASM emitter is by the x86 ASM language).

The C backend was not affected by this at all, because this optimization
checks for IR_OPT_ADDRCISC which is only used by the ASM backend. And also,
the C backend does not use the IDX/VREG mult field at all - because it is
only used with this specific optimization.